💄 style : Modify profile page Category.tsx design
