+++ src/pages/profile.tsx
@@ -1,15 +1,23 @@
-import {Box, Button, Card, CardContent, Grid, Skeleton, Tab, Tabs, Typography} from "@mui/material";
+import {Box, Button, Grid, Skeleton, Stack, Tab, Tabs, Typography} from "@mui/material";
 import { useNavigate } from "react-router-dom";
 import { useEffect, useState } from "react";
 import axios from "axios";
 import EditModal from "../components/profile/modals/EditModal";
 import ProfileImage from '../components/profile/information/ProfileImage';
 import Counter from "../components/profile/information/Counter";
-import MyButton from "../components/common/button/ProfileButton";
 import TabPanel from "../components/profile/tabs/TabPanel.tsx";
 import ReviewContainer from "../components/profile/reviewtab/ReviewContainer.tsx";
 import UserMap from "../components/profile/map/UserMap.tsx";
-import Categories from "../components/profile/category/Categories.tsx";
+import '../styles/profilepage.css';
+import AddStoryModal from "../components/profile/modals/AddStoryModal.tsx";
+import Category from "../components/profile/category/Category.tsx";
+import StoryModal from "../components/profile/modals/StoryModal.tsx";
+
+interface openModalProps {
+  id :number,
+  listName: string,
+  image: string,
+}
 
 function ProfilePage() {
 
@@ -26,7 +34,36 @@ function ProfilePage() {
   const queryString = window.location.search;
   const url = new URLSearchParams(queryString);
   const navigate = useNavigate();
+  const [storyList, setStoryList] = useState([]);
 
+  const [listModal, setListModal] = useState(false);
+
+  const [listInfo, setListInfo] = useState({
+    'id' : 0,
+    'name' : '',
+    'image' : '',
+  });
+
+  const openListModal = ({id, listName, image}: openModalProps) => {
+    setListModal(true);
+    setListInfo({
+      'id' : id,
+      'name' : listName,
+      'image' : image,
+    });
+  }
+
+  const closeListModal = () => {
+    setListModal(false);
+  }
+
+  const [addModal, setAddModal] = useState(false);
+  const openAddModal = () => {
+    setAddModal(true);
+  }
+  const closeAddModal = () => {
+    setAddModal(false);
+  }
   const [following, setFollowing] = useState(false);
   const [followAndReview, setFollowAndReview] = useState({
     reviews : 0,
@@ -87,6 +124,24 @@ function ProfilePage() {
     }
   }
 
+  useEffect(() => {
+    axios.get(`${import.meta.env.VITE_REACT_APP_BASE_URI}/api/storylist`,{
+      headers : {
+        Authorization : window.localStorage.getItem('access_token')
+      },
+      params: {
+        user_id: window.localStorage.getItem('id'),
+      }
+    })
+      .then(response => {
+        setStoryList(response.data);
+        console.log(response.data);
+      })
+      .catch(error => {
+        console.error(error);
+      });
+  },[addModal]);
+
   // Effect hook for get user data
   useEffect(() => {
     if(!window.localStorage.getItem('access_token')) {
@@ -105,7 +160,7 @@ function ProfilePage() {
       .then((response) => {
         const {id, nickname, profile_image, bio} = response.data;
         setUserData({id, nickname, profile_image, bio});
-        console.log('get')
+        console.log(response.data);
         setFollowAndReview({
           reviews : response.data.reviews,
           followers : response.data.follower,
@@ -155,53 +210,65 @@ function ProfilePage() {
   return (
       <Box>
         <EditModal userId={userData.id||'justin010129@gmail.com'} open={modalState} onClose={closeModal} />
-        <Grid container sx={{justifyContent: "center", marginTop: "3%", '@media (max-width: 770px)': {
+        <AddStoryModal userId={window.localStorage.getItem('id')||'justin010129@gmail.com'} open={addModal} onClose={closeAddModal} />
+        <Grid container sx={{justifyContent: "center", marginTop: "3%", '@media (max-width: 450px)': {
             flexDirection: 'column',
             alignItems: "center",
           }}}>
-          <Grid item xs={3} sx={{mr: "5%"}}>
-            <Card sx={{textAlign: "center", height: "fit-content", width: "fit-content", padding: "7%"}}>
-              <CardContent sx={{display: "flex", flexDirection :"column",
-                gap: "1", height: "fit-content",
-                '@media (max-width: 770px)': {
-                  flexDirection: 'row',
-                },}}>
-                <Box sx={{display: "flex", justifyContent: "center"}}>
-                  <Box sx={{position: "relative", paddingTop: "75%", width: "75%", height: "0"}}>
-                    <ProfileImage src={userData.profile_image} alt="ProfileImage" />
+          <Grid item xs={3}>
+            <Box className="profile-card">
+              <Box sx={{display: "flex", justifyContent: "center"}}>
+                <Box sx={{position: "relative", paddingTop: "60%", width: "60%", height: "0"}}>
+                  <ProfileImage src={userData.profile_image} alt="ProfileImage" />
+                </Box>
+              </Box>
+              {userData.nickname ?
+                  <Typography sx={{ fontSize : "1.1vw"}} variant="h5">{userData.nickname}</Typography> :
+                  <Skeleton sx={{mr: "10px"}} variant="rounded" width={50} height={20} />
+              }
+              {userData.bio ?
+                  <Typography variant="caption" sx={{my: "7px", fontSize: "0.9vw"}}>{userData.bio}</Typography> :
+                  <Skeleton sx={{mr: "10px"}} variant="rounded" width={50} height={18} />
+              }
+              {userData.id == window.localStorage.getItem('id') ?
+                  //   내 프로필 페이지일 경우
+                  <Box sx={{display: "flex", gap :"6", justifyContent: "space-evenly", flexWrap: "wrap"}}>
+                    <button className="profile-button" onClick={openModal} style={{ backgroundColor: "rgba(255, 164, 27, 0.85)" }}>編集</button>
+                    <button className="profile-button" onClick={destroyAccount} style={{ backgroundColor: "rgba(221, 0, 0, 0.85)" }}>退会</button>
                   </Box>
-                </Box>
-                {userData.nickname ?
-                    <Typography sx={{ fontSize : "1.1vw"}} component="span" variant="h5">{userData.nickname}</Typography> :
-                    <Skeleton sx={{mr: "10px"}} variant="rounded" width={50} height={20} />
-                }
-                {userData.bio ?
-                    <Typography variant="caption" sx={{my: "7px", fontSize: "0.8vw"}}>{userData.bio}</Typography> :
-                    <Skeleton sx={{mr: "10px"}} variant="rounded" width={50} height={18} />
-                }
-                {userData.id == window.localStorage.getItem('id') ?
-                    //   내 프로필 페이지일 경우
-                    <Box sx={{display: "flex"}}>
-                      <MyButton size={"small"} onClick={openModal} sx={{fontSize: "0.85vw", ":hover": {backgroundColor: "rgba(255, 164, 27, 1)", transform: "scale(1.01)", }, mx: "5px", backgroundColor: 'rgba(255, 164, 27, 0.85)', color: "black", borderRadius: "10%"}}>編集</MyButton>
-                      <MyButton size={"small"} onClick={destroyAccount} sx={{fontSize: "0.85vw", ":hover": {backgroundColor: "rgba(221, 0, 0)", transform: "scale(1.01)",}, mx: "5px", backgroundColor: "rgba(221, 0, 0, 0.85)", color: "black", borderRadius: "10%"}}> アカウント削除 </MyButton>
-                    </Box>
-                    :
-                    following ?
-                        <Button id={userData.id||"justin010129@gmail.com"} onClick={deleteFollow} variant="contained" size="small" sx={statusButtonStyle}>
-                          ファロー中
-                        </Button> :
-                        <Button id={userData.id||"justin010129@gmail.com"} onClick={addFollow} variant="contained" size="small" sx={statusButtonStyle}>
-                          ファローする
-                        </Button>
-                }
-                <Box sx={{ display: "flex", my: "8px", justifyContent: "center"}}>
-                  <Counter userId={userData.id||''} counterType="reviews" title="ポスト" count={followAndReview.reviews}/>
-                  <Counter userId={userData.id||''} counterType="follower" title="ファロワー" count={followAndReview.followers}/>
-                  <Counter userId={userData.id||''} counterType="following" title="ファロー中" count={followAndReview.followings}/>
-                </Box>
-              </CardContent>
-            </Card>
-
+                  :
+                  following ?
+                      <Button id={userData.id||"justin010129@gmail.com"} onClick={deleteFollow} variant="contained" size="small" sx={statusButtonStyle}>
+                        ファロー中
+                      </Button> :
+                      <Button id={userData.id||"justin010129@gmail.com"} onClick={addFollow} variant="contained" size="small" sx={statusButtonStyle}>
+                        ファローする
+                      </Button>
+              }
+              <Box sx={{ display: "flex", my: "8px", justifyContent: "center"}}>
+                <Counter userId={userData.id||''} counterType="reviews" title="ポスト" count={followAndReview.reviews}/>
+                <Counter userId={userData.id||''} counterType="follower" title="ファロワー" count={followAndReview.followers}/>
+                <Counter userId={userData.id||''} counterType="following" title="ファロー中" count={followAndReview.followings}/>
+              </Box>
+              <button className="profile-button" onClick={openAddModal}>レビューリスト作成</button>
+              <Stack direction="row" sx={{ my: "5px", justifyContent : "left"}} flexWrap="wrap">
+                {storyList.map((list) => (
+                  <Category key={list['id']} id={list['id']} onClick={()=>{openListModal({
+                    id : list['id'],
+                    listName : list['story_name'],
+                    image : list['mainImage']})}} src={list['mainImage']} title={list['story_name']}/>
+                ))}
+              </Stack>
+              {listModal && (
+                <StoryModal
+                  open={listModal}
+                  onClose={closeListModal}
+                  id={listInfo.id}
+                  image = {listInfo.image}
+                  storyName= {listInfo.name}
+                />
+              )}
+            </Box>
           </Grid>
           <Grid item xs={7}>
             <Tabs variant="fullWidth" value={tab} onChange={selectTab} aria-label="usertabs" sx={{
@@ -222,19 +289,13 @@ function ProfilePage() {
                   color: "black",
                 },
               }}/>
-              <Tab label="My Logs" value={3} sx={{
-                "&.Mui-selected": {
-                  color: "black",
-                },
-              }}/>
             </Tabs>
 
-            <TabPanel value={tab} index={1}><ReviewContainer userId={userData.id||""} /></TabPanel>
+            <TabPanel value={tab} index={1}>
+              <ReviewContainer userId={userData.id||""} />
+            </TabPanel>
             <TabPanel value={tab} index={2}>
               <UserMap userId={userData.id||""} />
-            </TabPanel>
-            <TabPanel value={tab} index={3}>
-              <Categories id={userData.id||""} />
             </TabPanel>
           </Grid>
         </Grid>
